refactor: move preview code to module
