+++ lua/gitsigns/diffthis.lua
@@ -49,11 +49,11 @@ local function bufread(bufnr, dbufnr, base)
   require('gitsigns.attach').attach(dbufnr, nil, 'BufReadCmd')
 end
 
+--- @async
 --- @param bufnr integer
 --- @param dbufnr integer
 --- @param base string?
---- @param _callback? fun()
-local bufwrite = async.async(function(bufnr, dbufnr, base, _callback)
+local function bufwrite(bufnr, dbufnr, base)
   local bcache = assert(cache[bufnr])
   local buftext = util.buf_lines(dbufnr)
   base = util.norm_base(base)
@@ -69,7 +69,7 @@ local bufwrite = async.async(function(bufnr, dbufnr, base, _callback)
     bcache.compare_text = buftext
     manager.update(bufnr)
   end
-end)
+end
 
 --- @async
 --- Create a gitsigns buffer for a certain revision of a file
@@ -114,7 +114,7 @@ local function create_revision_buf(bufnr, base)
       group = 'gitsigns',
       buffer = dbuf,
       callback = function()
-        bufwrite(bufnr, dbuf, base)
+        async.arun(bufwrite, bufnr, dbuf, base)
       end,
     })
   else
@@ -252,10 +252,11 @@ local function is_fugitive_diff_window(name)
     and vim.fn.FugitiveParse(name)[1] ~= ':'
 end
 
--- This function needs to be throttled as there is a call to vim.ui.input
+--- This function needs to be throttled as there is a call to vim.ui.input
+--- @async
 --- @param bufnr integer
 --- @param _callback? fun()
-M.update = throttle_by_id(async.create(1, function(bufnr, _callback)
+M.update = throttle_by_id(function(bufnr, _callback)
   if not vim.wo.diff then
     return
   end
@@ -278,6 +279,6 @@ M.update = throttle_by_id(async.create(1, function(bufnr, _callback)
       end
     end
   end
-end))
+end)
 
 return M
